p08: Commit working solution.
I changed the algorithm a bit, which runs much faster. The new algorithm
initially had the same wrong answer.

Bug seems to have been that I was too cautious on the check to avoid
adding cards beyond the end of the table. It worked for the sample data
but the real data was more aggressive.