p17: Fix Perl-based attempt.
The issue was with the check to terminate. By summing all values in the
list instead of checking for equality with 0, we could accidentally
permit list like [-2  0 2] to terminate the loop early.

summing absolute values also fixed this but I prefer to have a check
closer to the source requirement.