p23: Commit working solution.
My generation logic was flawed, in that it was only possible to have one
interstitial space be larger than 1 wide.

Since I knew there had to be a combinatorial defined solution to the
number of blocks in this situation from working with nonograms, I
googled the closed-form solution and used that to find failing test
cases (extracted into other sample files).

From there I had to figure out how to count all the possible
interstitial solutions and generate valid solutions.  Took me awhile but
it worked.