p24: Commit rewrite that still works on part 1.
Brute force is literally impossible with the scale of some of these
solutions. Dynamic programming or memoized recursion were not possible
with the last approach, so rewrite to use recursion.

This works on the full part 1 input and the samples.  Now to add the
actual memoization so I can try it on part 2 input.