p33: Speed up search somewhat.
The last commit actually gave the correct answer after a few minutes,
but that's not going to work for whatever is in part 2.

There are a couple of correctness changes (e.g. consecutive steps from
the first cell), but the big change is to switch to a priority queue to
keep the list of nodes to visit sorted on the fly.

Switching to clang and libc++ also helps improve performance. However
it's still about 12 seconds to run on the full input.