p36: Commit working solution. Not thrilled.
Not thrilled because this was just a "you know the math formula for it
or you don't" problem. As soon as I saw what they wanted for the input I
just went to the solution thread.

There's no way I would have spent the time coming up with scanline or
rectangle-based approaches that work with infinite grids.

But whatever, the star shows up on my account now.