p34: Use vector rather than hash_table for node table.
Even on part 2 rules on the full input the memory usage is not very
noticeable on my machine, and this on top of the other optimizations
reduces the runtime for part 2 on the input from ~20m to something more
like 3m.  Still too long but further improvements are more likely to be
algorithmic than from tweaking data structures.