p34: Holy Grail! Got the lighter node type working and it is much faster.
0.31 seconds on the full input for part 2, which previously took 20
minutes in the passing attempt (and still many minutes up to this
point).

This seems mostly down to making the graph node for the Dijkstra
algorithm contain less information (position and next move orientation),
as suggested by some on the AoC subreddit.

By requiring every "visit" to consider every possible straight line move
configuration, you can ensure that the "transition" to the next cell
will always require a left or right turn when that cell is visited.

So we don't have to record either incoming direction or consecutive
steps. We don't even need direction at all, just whether a right/left
turn will have us going horizontally or vertically.

Because there are fewer possible nodes, there is less auxiliary
bookkeeping to keep correct. I'm not sure that's enough to explain the
different but the facts don't lie.

There remains weirdness about how to handle the start node that infects
some of the code but it's nothing too bad.