p34: Make a dedicated end node subtype.
This allows for there to be only a single actual end node, which allows
use of the Dijkstra's algorithm shortcut to stop searching as soon as
the end node pops up as a node to be visited, even if other nodes are
pending.

This in turn might make heuristic-based addons like A* more useful but
I'm not going that far.