2015/14: Implement solution for part 2.
Although this was extremely easy to cheese for the second star (just
edit the input...), I went on to do this legit.

A challenge is that the Op class basically has to cache intermediate
values as the performance is intractable otherwise, but it's hard to
reset those nodes after.

Global vars fix all here, but a more solid solution might be something
like a 'data version' maintained in the Puzzle. Doesn't matter now
though.