p41: Works on sample input.
We don't need to track steps after all. Dijkstra's normal behavior will
prevent searching visited nodes and there's no difference in outcome if
you land on a spot with N + 2 steps left or with N steps left.  So just
search every open spot until we run out of steps.

With this change, and s/break/continue/ in the neighbor loop, it passes
on sample input.