p42: Even more debugging output.
This makes the website happy. I won't say 'solved the puzzle', I just
went through the motions at this point because it turned into another
"the input is magic exercise" again.

The tutorial at https://github.com/villuna/aoc23/wiki/A-Geometric-solution-to-advent-of-code-2023,-day-21
was very useful here. I had parity down but wasn't going to map out all
the possible squares nearly as well as this was done.

I think I had already managed to get the various constants out of the
previous version of the code (by using different number of steps to
switch the parity of the puzzle while subdividing to 65) but
fat-fingered the earlier calculation.