p43: More bugfixes, still wrong.
This time the answer is 1384, which is still too high.

Primary bug was that it is basically mandatory to track all bricks in
both zx and zy orientations, so that you know which bricks to move down
or not.

Not sure what the issue is now or a good way to reproduce. The input is
much too large to do anything substantial with.