p43: Found another bug, added test. Still broken.
Added some blocks to the example data until I found another bug. The
block detection code was supposed to look for overlap in both zx and zy
planes, but it was buggy, revealed by brick 9 in the updated test case.

What it was actually doing was looking at all blocks at the lower z
level, and if it was supported by *any* block, marked *all* lower blocks
as a supporting block. Obviously this doesn't work for all geometries,
even if it happened to work for the example data.

That all said, things are going the wrong direction still. Now we get an
answer of 729, which is so far off that you don't even get a hint from
the website whether it's too high or too low (so, clearly too low).