p43: Do not overcount supporting blocks in zx/zy.
Bug caught by comparison of $(head -10 input) with working
implementation written in Python.

We're looking for a single block below to support the current block in
both the zx and zy planes but it's possible for a block to support in
2 or more of the zx plane and not at all in the zy. Use a saturating
check instead.

This gives an answer which is much closer but still slightly off.