2015/29: Working solution for Day 15 part 1.
Some complexity in trying to make a single enumerator that spits out all
combinations from a multi-level counter of base n.

Slow since it's brute force, but not so slow as to be worth dusting off
the ol' linear algebra textbook to figure out how to do this with basis
vectors and the like.