2024/04: Implement solution for Day 2, puzzle 2.
I was overthinking this at first thinking to make it more efficient by
looping over the list of numbers and just 'skipping' one if it would
result in an unsafe reading. But I realized this would make things more
difficult if the first number of the list (before a trend had been
established) could be removed.

Ended up doing the dumb simple thing (literally making a list of every
possible sub-set of one number being removed) and it worked first try,
and quickly at that.