2015/38: Implement working solution for day 19 part 2.
I don't like this because I feel like it's not logically sound in
general, as I can't prove that my heuristic of reversing the
longest-possible productions first is actually the quickest way to get
from the final string back to the source electron.

However the 'proper' Djikstra algorithm will take weeks if my computer
doesn't run out of memory first, and doing it this way *did* get me
my gold star... so here it is.