2024/12: Working implementation for day 6, part 2. First try.
Although I got it the first try, the initial code was slow. This
represents a bit of refactoring to use a hash set instead of a vector
when looking for cycles.

This is still slow by Rust standards but it is definitely good enough
for me.

This is the first one where I feel I really had to fight the
borrow-checker. Closures were easy to use, on the other hand.  Will be
interesting if it gets worse going forward (I'm sure it will!...).