2024/16: Significant simplification.
The exercise today doesn't actually involve anything grid-specific, it's
all about the antennae and their antipodes, the only thing that relates
back to the 2D workspace is the bounds-check.

So we can refactor the loading and bounds-check code a bit and get rid
of everything else, which significantly shortens the code.