2024/16: More simplifications.
It's possible to use some Rust-style iterators instead of a for loop to
generate the antipodes, which simplifies things significantly even at
the cost of a couple of aux functions.

With that change we don't need the hard-coded combination function, the
loops are easy to generate inline and now that the core loop is so small
it doesn't distract in the code to do this.