2015/47: Implement working solution for day 24 part 1. A few retries.
This is hellaciously slow but I wasn't able to prove to myself that all
possible combinations of remaining weights would add to the required
weight once we found a minimum-package-count weight for group 1.

Because of this I did a ton of preprocessing of the number of weights to
look for all possible combinations between min and max numbers of
weights that added up to the required amount.

I think encode that to a bitstring (supporting up to 32 weights, which
was within the size of the full puzzle input), to map an entire group of
weights to a single 32-bit work, where bit X being set meant that
weight[bit] was part of the group.

From there, basic binary operations like popcount and OR/XOR can be used
to figure out whether two groups are trying to vie for the same unique
weights or not.

My biggest problem was trying to use binary search to speedup the check
for the third group from the pre-calculated possible groups. Ruby's
bsearch wasn't working for some reason, probably my fault, but replacing
that lookup with find got me back on track.

Note that debugging this got me to peeking at other solutions, and it
turns out the input is extremely nice in that you will never have to
verify group 2 or 3 once you find group 1 the obvious way... would have
been nice to know, lol.