2024/28: Working solution for day 14, part 2.
I thought we could get away with finding a common multiple of some
variously-timed cycles, but it turned out there was a single cycle for
every robot of about width * height.

So after some light research I figured it was easier to search for
'weird' safety scores, and that helped in narrowing a filter in the code
to look for suspicious cases and print them out.  This gave the working
answer.