2024/{37,38}: Add Rust implementations of the Ruby solutions.
Oddly, I had sorted the list of towels thinking initially I would binary
search through the towels. I ended up going a different way in the Ruby
impl but left it sorted.

When I removed the sort from the Rust implementing thinking that it was
unnecessary, it broke the output for the full input.  Still not
entirely sure why...