2024/42: Working solution for day 21, part 2.
Continuing with Ruby for now (which was quite fun for this one!). First
try although it took me a worryingly long time to get the recursive bugs
worked out. I'm still not sure what specifically changes from the sample
being broken to working fine, but that version also worked on full
input.