2024/46: Implement working solution for day 23 part 2.
Several tries, the struggle was because of something I had briefly
considered and then forgot about. Initially I had my loop something like

loop all node-edge pairs {
  loop large to small combinations of edges from node {
    Look for match
  }
}

The issue is that you can find a match for a given node with a smaller
number of combinations than is possible in the entire input. So the
outer loop needs to be the size of the graph to consider so that you
guarantee any solutions of the highest size will be found first.

Aside from that it wasn't too bad, Ruby again makes some of the
arbitrary array size stuff easy.