2019/15: Working solution for day 8, part 1.
Not the first try but that's because I kept misreading the problem (min
vs max).

But I blame that partially on fighting through some QuickJS bugs
relating to generator support, which apparently does not fully implement
iterator support.  Not sure what the point is to generators if you have
to bake them into a materialized array anyways but maybe I'll figure out
something else later.