2019/20: Working solution for day 10, part 2. First try.
This was only workable because my thought that the atan2 trick to
reverse y and x might be useful in fixing the rotation and orientation
simultaneously to match the problem to the existing atan2 ended up
paying off on the first try.

That made the rest a fairly straightforward exercise of mapping
asteroids to trigonometric angles, sorting by distance and/or angle as
appropriate, and then just cycling through the result until everything
was accounted for.