2019/26: Implement working solution for day 13, part 2. First try.
The hard part is that the script is synchronous JavaScript and this
really plays against what NodeJS wants to do (for good reason).

So there's hacks like spawning external child processes to sleep, or
reading directly from fd 1 in some commented code from when the input
was based on user input for testing instead of my program generating the
right input.

But it did work and will get me onto the next puzzle.