2025/02: Working solution for Day 1 part 2.
Again not quite the first try, I ended up with a much dumber solution to
get the star (literally turn the lock one-by-one and count zeroes), and
then used that to as a bridge to get to this.

Impressive puzzle, really makes you pay attention to detail here, not
bad for a day 1 this year!