2025/11: Working solution for day 6, part 1. First try.
As I suspected from seeing the problem, the hardest part by far was
getting the input handling right, lol.

Once I got the number read into the lists it was quite easy to do the
reductions (I do want to go back and view/rangify it but just wanted the
star for now).