2025/14: Working solution for day 7, part 2. First try again.
On this one you have to be very careful that you know what's being added
or not, so I ended up documenting in sample2 the exact configuration
that gets you to 40 timelines on the sample problem to make sure I
understood the login.

Once that was done, it was straightforward to convert part 1 to part 2.