2025/18: Working solution for day 9, part 2.
This one was pretty miserable, truth be told. As I suspected, there was
a good deal of pain here and even though I had the general algorithm in
mind from early on, getting all the details right was a huge challenge.
Eventually picked up an easier-to-implement shortcut but even that was a
slog in a sea of > and <= and whatnot.

4 tries total, 0 fun gained.