2025/19: Working solution for day 10, part 1.
Second try, but wasn't too bad to correct my Djikstra implementation to
get the answer on the second try.

Oddly, the code gave correct answers only in debug mode. I needed a bit
of an AI assist to point me to my Toggles array in decode_input_line
that had been initially uninitialized. Debug mode cleared it to 0 but
optimized mode did not and that caused other problems.

Part 2 is going to be a crapshoot, needing matrix math to address
joltage requirements. We'll see if I do it or not, linear algebra is not
my preferred way of spending Christmas, and there's no 'puzzle' it's
just implementation...